% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_experiment_db}
\alias{filter_experiment_db}
\title{Apply filter to experiment database}
\usage{
filter_experiment_db(filters, experiment_db = "experiment_database.csv")
}
\arguments{
\item{filters}{string specifying the filter conditions}

\item{experiment_db}{experiment database file}
}
\value{
data.frame containing the filtered experiment database
}
\description{
filter the experiment database by the available columns
}
