% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_experiments}
\alias{filter_experiments}
\title{Filter experiments without constructing a database}
\usage{
filter_experiments(filters, path = "*/*/data/*/metadata")
}
\arguments{
\item{filters}{string specifying the filter conditions}

\item{path}{character vector of patterns for relative or absolute filepaths. Missing values will be ignored.}
}
\value{
data.frame containing a filtered experiment database
}
\description{
Filter experiments without constructing a database
}
