% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_filtered_data}
\alias{get_filtered_data}
\title{Get filtered experimental data}
\usage{
get_filtered_data(filtered_locations)
}
\arguments{
\item{filtered_locations}{data.frame returned by filter_experiment_db or filter_experiments}
}
\value{
data.frame containing filtered experimental data
}
\description{
Get filtered experimental data
}
