% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_experiment_db}
\alias{make_experiment_db}
\title{Make experiment database}
\usage{
make_experiment_db(path = "*/*/data/*/metadata")
}
\arguments{
\item{path}{character vector of patterns for relative or absolute filepaths.
Missing values will be ignored.}
}
\description{
Compile plate layout .csv files into a single .csv file to use for experiment
filtering. Writes a .csv file to one directory up from the first wildcard in
the given path.
}
